+++ Assets/PersonalWorkplace/BIK/Prefabs/MainSceneCanvas.prefab
@@ -1290,7 +1290,7 @@ MonoBehaviour:
   - type: 7
     ui: {fileID: 2748968418779302911}
   - type: 11
-    ui: {fileID: 1629872783396112511}
+    ui: {fileID: 1897682459508885515}
   - type: 14
     ui: {fileID: 7934852552561887357}
   _heroUI: {fileID: 5706282459793846866}
@@ -5141,15 +5141,15 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: a4ef8ebe1ca5c9a468e6c3662d53cffd, type: 3}
---- !u!114 &1629872783396112511 stripped
+--- !u!114 &1897682459508885515 stripped
 MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5229119862277367178, guid: a4ef8ebe1ca5c9a468e6c3662d53cffd, type: 3}
+  m_CorrespondingSourceObject: {fileID: 4925253824032548350, guid: a4ef8ebe1ca5c9a468e6c3662d53cffd, type: 3}
   m_PrefabInstance: {fileID: 6777891491107597813}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4acf81218f2b6cc4195a1d26df7c95c0, type: 3}
+  m_Script: {fileID: 11500000, guid: 36a66d6f1e84d5c409fb624beecd8e54, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!224 &3401601699971634130 stripped
