+++ Assets/PersonalWorkplace/MSK/MSK_Scene/MSK_DEMO_GameScene.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b7eb2b534c235df489234a383b5b8084
+guid: bb090509d5b713c4c942a7765ed2c27e
 DefaultImporter:
   externalObjects: {}
   userData:
