+++ Assets/PersonalWorkplace/MSK/MSK_Scripts/UI/SummonUI.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 4acf81218f2b6cc4195a1d26df7c95c0
+guid: 36a66d6f1e84d5c409fb624beecd8e54
